update project name to be capitalized, changed page names, update styles to align with best practices and uses light/dark mode
