+++ Samples/SnowPal-Win-101/SnowPal.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35931.197 d17.13
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SnowPal", "SnowPal\SnowPal.csproj", "{74C93DF2-EEA7-4C65-881C-DA1590573D5F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|ARM64 = Debug|ARM64
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|ARM64 = Release|ARM64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|ARM64.Build.0 = Debug|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x64.ActiveCfg = Debug|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x64.Build.0 = Debug|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x64.Deploy.0 = Debug|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x86.ActiveCfg = Debug|x86
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x86.Build.0 = Debug|x86
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Debug|x86.Deploy.0 = Debug|x86
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|ARM64.ActiveCfg = Release|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|ARM64.Build.0 = Release|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|ARM64.Deploy.0 = Release|ARM64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x64.ActiveCfg = Release|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x64.Build.0 = Release|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x64.Deploy.0 = Release|x64
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x86.ActiveCfg = Release|x86
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x86.Build.0 = Release|x86
+		{74C93DF2-EEA7-4C65-881C-DA1590573D5F}.Release|x86.Deploy.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {20800B6A-41A9-4EF4-88C5-E59AAB98015E}
+	EndGlobalSection
+EndGlobal
