+++ Samples/SnowPal-Win-101/Snowpal/App.xaml.cs
