+++ Samples/SnowPal-Win-101/Snowpal/Assets/SplashScreen.scale-200.png
