+++ Samples/SnowPal-Win-101/Snowpal/Assets/Square150x150Logo.scale-200.png
