+++ Samples/SnowPal-Win-101/Snowpal/Assets/Square44x44Logo.scale-200.png
