+++ Samples/SnowPal-Win-101/Snowpal/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
