+++ Samples/SnowPal-Win-101/Snowpal/Assets/Wide310x150Logo.scale-200.png
