+++ Samples/SnowPal-Win-101/Snowpal/MainWindow.xaml
@@ -7,7 +7,8 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
-    Title="snowpal">
-    
+    Title="SnowPal">
+
     <Frame x:Name="MainFrame" />
+
 </Window>
