+++ Samples/SnowPal-Win-101/Snowpal/Package.appxmanifest
@@ -8,14 +8,14 @@
   IgnorableNamespaces="uap rescap">
 
   <Identity
-    Name="428c1c1b-84f4-407f-b23a-fe35f63f972e"
+    Name="a400bf75-bbf4-4376-ae4d-3d52ecd0f382"
     Publisher="CN=jearleycha"
     Version="1.0.0.0" />
 
-  <mp:PhoneIdentity PhoneProductId="428c1c1b-84f4-407f-b23a-fe35f63f972e" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
+  <mp:PhoneIdentity PhoneProductId="a400bf75-bbf4-4376-ae4d-3d52ecd0f382" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
 
   <Properties>
-    <DisplayName>snowpal</DisplayName>
+    <DisplayName>SnowPal</DisplayName>
     <PublisherDisplayName>jearleycha</PublisherDisplayName>
     <Logo>Assets\StoreLogo.png</Logo>
   </Properties>
@@ -34,8 +34,8 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="snowpal"
-        Description="snowpal"
+        DisplayName="SnowPal"
+        Description="SnowPal"
         BackgroundColor="transparent"
         Square150x150Logo="Assets\Square150x150Logo.png"
         Square44x44Logo="Assets\Square44x44Logo.png">
