+++ Samples/SnowPal-Win-101/Snowpal/Pages/MenuPage.xaml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="SnowPal.Pages.MenuPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:SnowPal.Pages"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Page.Background>
+        <LinearGradientBrush StartPoint="0,0" EndPoint="1,1">
+            <GradientStop Offset="0.0" Color="{StaticResource GradientStop1}" />
+            <GradientStop Offset="0.5" Color="{StaticResource GradientStop2}" />
+            <GradientStop Offset="1.0" Color="{StaticResource GradientStop3}" />
+        </LinearGradientBrush>
+    </Page.Background>
+
+    <Grid>
+
+        <!-- Title -->
+        <TextBlock Text="SnowPal" Style="{StaticResource TitleTextStyle}" />
+
+        <!-- Menu Buttons -->
+        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Spacing="20">
+            <Button Content="New Game" Style="{StaticResource MenuButton}" Click="NewGame_Click"/>
+            <Button Content="Leaderboard" Style="{StaticResource MenuButton}" Click="Leaderboard_Click"/>
+            <Button Content="Options" Style="{StaticResource MenuButton}" Click="Options_Click"/>
+            <Button Content="Credits" Style="{StaticResource MenuButton}" Click="Credits_Click"/>
+            <Button Content="Exit" Style="{StaticResource MenuButton}" Click="Exit_Click"/>
+        </StackPanel>
+    </Grid>
+</Page>
