+++ Samples/SnowPal-Win-101/Snowpal/Pages/MenuPage.xaml.cs
@@ -1,13 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Controls.Primitives;
@@ -15,28 +5,32 @@ using Microsoft.UI.Xaml.Data;
 using Microsoft.UI.Xaml.Input;
 using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
-using SnowPal;
-
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices.WindowsRuntime;
+using Windows.Foundation;
+using Windows.Foundation.Collections;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
 
-namespace SnowPal
+namespace SnowPal.Pages
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
     /// </summary>
-    public sealed partial class StartPage : Page
+    public sealed partial class MenuPage : Page
     {
-        public StartPage()
+        public MenuPage()
         {
             this.InitializeComponent();
         }
-
         private void NewGame_Click(object sender, RoutedEventArgs e)
         {
-            // Navigate to the MainPage
-            Frame.Navigate(typeof(MainPage));
+            // Navigate to the GamePage
+            Frame.Navigate(typeof(GamePage));
         }
 
         private void Leaderboard_Click(object sender, RoutedEventArgs e)
