+++ Samples/SnowPal-Win-101/Snowpal/Properties/launchSettings.json
@@ -1,9 +1,9 @@
 {
   "profiles": {
-    "snowpal (Package)": {
+    "SnowPal (Package)": {
       "commandName": "MsixPackage"
     },
-    "snowpal (Unpackaged)": {
+    "SnowPal (Unpackaged)": {
       "commandName": "Project"
     }
   }
