+++ Samples/SnowPal-Win-101/Snowpal/Styles.xaml
@@ -2,100 +2,81 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    <!-- Define a base style for Title TextBlock -->
-    <Style x:Key="BaseTextBlockStyle" TargetType="TextBlock">
+    <!-- Define a style for Title TextBlock -->
+    <Style x:Key="TitleTextStyle" TargetType="TextBlock">
         <Setter Property="FontFamily" Value="Segoe UI"/>
-        <Setter Property="Foreground" Value="Black"/>
         <Setter Property="FontSize" Value="48"/>
         <Setter Property="FontWeight" Value="Bold"/>
         <Setter Property="HorizontalAlignment" Value="Center"/>
         <Setter Property="VerticalAlignment" Value="Top"/>
         <Setter Property="Margin" Value="0,50,0,0"/>
     </Style>
- 
-    <!-- Define a style for the Title TextBlock that inherits from BaseTextBlockStyle -->
-    <Style x:Key="MenuTitleTextBlockStyle" TargetType="TextBlock" BasedOn="{StaticResource BaseTextBlockStyle}">
-        <Setter Property="Foreground" Value="White"/>
-    </Style>
 
-    <!-- Define a style for the Menu Buttons -->
-    <Style x:Key="MenuButtonStyle" TargetType="Button">
-        <Setter Property="Width" Value="200"/>
-        <Setter Property="Height" Value="50"/>
-        <Setter Property="Margin" Value="2"/>
+    <!-- BUTTONS -->
+    <!-- Define a style for Bass Buttons -->
+    <Style x:Key="BaseButton" TargetType="Button">
         <Setter Property="FontFamily" Value="Segoe UI"/>
         <Setter Property="FontSize" Value="20"/>
-    </Style>
-
-    <!-- Define a style for the Guesses Left TextBlock -->
-    <Style x:Key="GuessesLeftTextBlockStyle" TargetType="TextBlock">
-        <Setter Property="FontSize" Value="20"/>
+        <Setter Property="Margin" Value="2"/>
     </Style>
 
     <!-- Define a style for Alphabet Buttons -->
-    <Style x:Key="AlphabetButtonStyle" TargetType="Button">
+    <Style x:Key="AlphabetButton" TargetType="Button" BasedOn="{StaticResource BaseButton}">
         <Setter Property="Width" Value="48"/>
         <Setter Property="Height" Value="48"/>
-        <Setter Property="Margin" Value="2"/>
-        <Setter Property="FontFamily" Value="Segoe UI"/>
         <Setter Property="FontSize" Value="25"/>
+    </Style>
+    <!-- Define a style for the Menu Buttons -->
+    <Style x:Key="MenuButton" TargetType="Button" BasedOn="{StaticResource BaseButton}">
+        <Setter Property="Width" Value="200"/>
+        <Setter Property="Height" Value="50"/>
+    </Style>
+    <!-- Define a style for the Popup Buttons -->
+    <Style x:Key="PopupButton" TargetType="Button" BasedOn="{StaticResource BaseButton}">
+        <Setter Property="HorizontalAlignment" Value="Right"/>
+    </Style>
+
+    <!-- TextBlock -->
+    <!-- Define a style for TextBlock -->
+    <Style x:Key="BaseTextBlock" TargetType="TextBlock">
+        <Setter Property="FontSize" Value="24"/>
+        <Setter Property="TextWrapping" Value="Wrap"/>
     </Style>
 
     <!-- Define a style for the Word Display TextBlock -->
-    <Style x:Key="WordDisplayTextBlockStyle" TargetType="TextBlock">
-        <Setter Property="FontSize" Value="50"/>
+    <Style x:Key="LargerTextBlock" TargetType="TextBlock" BasedOn="{StaticResource BaseTextBlock}">
+        <Setter Property="FontSize" Value="40"/>
         <Setter Property="HorizontalAlignment" Value="Center"/>
         <Setter Property="VerticalAlignment" Value="Bottom"/>
         <Setter Property="Margin" Value="10"/>
     </Style>
-    
-    <!-- Define a style for the Menu Buttons -->
-    <Style x:Key="MenuButtons" TargetType="Button">
-        <Setter Property="Width" Value="200"/>
-        <Setter Property="Height" Value="50"/>
-    </Style>
 
-
+    <!-- Popup -->
     <!-- Define a style for the Popup -->
     <Style x:Key="PopupStyle" TargetType="Popup">
         <Setter Property="VerticalOffset" Value="300"/>
         <Setter Property="HorizontalOffset" Value="800"/>
     </Style>
-    
-    <!-- Define a style for the MessageTitle Popup TextBlockStyle -->
-    <Style x:Key="MessageTitlePopupTextBlockStyle" TargetType="TextBlock">
-        <Setter Property="FontSize" Value="50"/>
-        <Setter Property="HorizontalAlignment" Value="Center"/>
-    </Style>
- 
 
-    <!-- Define a style for MessageContent Popup TextBlockStyle -->
-    <Style x:Key="MessageContentPopupTextBlockStyle" TargetType="TextBlock">
-        <Setter Property="FontSize" Value="24"/>
-        <Setter Property="TextWrapping" Value="Wrap"/>
-    </Style>
 
-    <!-- Define a style for the Popup Button -->
-    <Style x:Key="PopupButtonStyle" TargetType="Button">
-        <Setter Property="HorizontalAlignment" Value="Right"/>
-    </Style>
-
+    <!-- StackPanel -->
     <!-- Define a base Center style for StackPanel  -->
-    <Style x:Key="HorizontalCenterStackPanel" TargetType="StackPanel">
+    <Style x:Key="CenterStackPanel" TargetType="StackPanel">
         <Setter Property="HorizontalAlignment" Value="Center"/>
-    </Style>
-
-    <Style x:Key="TopHorizontalCenterStackPanel" TargetType="StackPanel" BasedOn="{StaticResource HorizontalCenterStackPanel}">
-        <Setter Property="Orientation" Value="Horizontal"/>
-        <Setter Property="VerticalAlignment" Value="Top"/>
-        <Setter Property="Margin" Value="10"/>
-    </Style>
-
-    <Style x:Key="CenterHorizontalCenterStackPanel" TargetType="StackPanel" BasedOn="{StaticResource HorizontalCenterStackPanel}">
         <Setter Property="VerticalAlignment" Value="Center"/>
         <Setter Property="Spacing" Value="12"/>
     </Style>
-    
+    <!-- Define a Horizontal for StackPanel  -->
+    <Style x:Key="HorizontalCenterStackPanel" TargetType="StackPanel" BasedOn="{StaticResource CenterStackPanel}">
+        <Setter Property="Orientation" Value="Horizontal"/>
+    </Style>
+    <!-- Define a Vertical for StackPanel  -->
+    <Style x:Key="VerticalCenterStackPanel" TargetType="StackPanel" BasedOn="{StaticResource CenterStackPanel}">
+        <Setter Property="Orientation" Value="Vertical"/>
+    </Style>
+
+
+    <!-- Image -->
     <!-- Define style for Image -->
     <Style x:Key="ImageStyle" TargetType="Image">
         <Setter Property="HorizontalAlignment" Value="Center"/>
@@ -103,19 +84,21 @@
         <Setter Property="Margin" Value="10,10,10,10"/>
     </Style>
 
+
+    <!-- GridView -->
     <!-- Define style for GridView -->
     <Style x:Key="GridViewStyle" TargetType="GridView">
         <Setter Property="HorizontalAlignment" Value="Center"/>
         <Setter Property="VerticalAlignment" Value="Center"/>
         <Setter Property="Margin" Value="10,50,10,10"/>
     </Style>
-
     <!-- Define style for GridView -->
     <Style x:Key="GridStyle" TargetType="Grid">
         <Setter Property="HorizontalAlignment" Value="Left"/>
         <Setter Property="VerticalAlignment" Value="Top"/>
     </Style>
 
+    <!-- Border -->
     <!-- Define style for Border -->
     <Style x:Key="BorderStyle" TargetType="Border">
         <Setter Property="Padding" Value="20"/>
@@ -124,6 +107,24 @@
         <Setter Property="Height" Value="400"/>
         <Setter Property="BorderThickness" Value="2"/>
     </Style>
+
+    <!-- Colors -->
+    <!-- Theme Dictionaries -->
+    <ResourceDictionary.ThemeDictionaries>
+        <!-- Light Theme -->
+        <ResourceDictionary x:Key="Light">
+            <Color x:Key="GradientStop1">#FFFFFF</Color>
+            <Color x:Key="GradientStop2">#ADD8E6</Color>
+            <Color x:Key="GradientStop3">#C9E4CA</Color>
+        </ResourceDictionary>
+
+        <!-- Dark Theme -->
+        <ResourceDictionary x:Key="Dark">
+            <Color x:Key="GradientStop1">#1E1E1E</Color>
+            <Color x:Key="GradientStop2">#2A2A72</Color>
+            <Color x:Key="GradientStop3">#3A3A3A</Color>
+        </ResourceDictionary>
+    </ResourceDictionary.ThemeDictionaries>
 
 
 </ResourceDictionary>
