--- Samples/SnowPal-Win-101/snowpal.sln
@@ -1,40 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "snowpal", "snowpal\snowpal.csproj", "{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|ARM64.Build.0 = Debug|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x64.ActiveCfg = Debug|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x64.Build.0 = Debug|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x64.Deploy.0 = Debug|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x86.ActiveCfg = Debug|x86
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x86.Build.0 = Debug|x86
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Debug|x86.Deploy.0 = Debug|x86
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|ARM64.ActiveCfg = Release|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|ARM64.Build.0 = Release|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|ARM64.Deploy.0 = Release|ARM64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x64.ActiveCfg = Release|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x64.Build.0 = Release|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x64.Deploy.0 = Release|x64
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x86.ActiveCfg = Release|x86
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x86.Build.0 = Release|x86
-		{5C6F1CA2-CB59-4EC4-9462-E6D73DA04F83}.Release|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
