--- Samples/SnowPal-Win-101/snowpal/MainWindow.xaml.cs
@@ -1,22 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Reflection;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Input;
-using SnowPal;
-
-// To learn more about WinUI, the WinUI project structure,
-// and more about our project templates, see: http://aka.ms/winui-project-info.
-
-namespace SnowPal
-{
-    public sealed partial class MainWindow : Window
-    {
-        public MainWindow()
-        {
-            this.InitializeComponent();
-            MainFrame.Navigate(typeof(StartPage));
-        }
-    }
-}
