--- Samples/SnowPal-Win-101/snowpal/Pages/MainPage.xaml.cs
@@ -1,16 +1,0 @@
-using Microsoft.UI.Xaml.Controls;
-
-
-
-namespace SnowPal
-{
-    public sealed partial class MainPage : Page
-    {
-        public SnowpalViewModel ViewModel { get; } = new();
-
-        public MainPage()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
