--- Samples/SnowPal-Win-101/snowpal/Pages/StartPage.xaml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Page
-    x:Class="SnowPal.StartPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:SnowPal"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Page.Background>
-        <LinearGradientBrush StartPoint="0,0" EndPoint="1,1">
-            <GradientStop Offset="0.0" Color="#FFFFFF" />
-            <GradientStop Offset="0.5" Color="#ADD8E6" />
-            <GradientStop Offset="1.0" Color="#C9E4CA" />
-        </LinearGradientBrush>
-    </Page.Background>
-
-    <Grid>
-
-        <!-- Title -->
-        <TextBlock Text="SnowPal" Style="{StaticResource MenuTitleTextBlockStyle}"/>
-
-        <!-- Menu Buttons --> 
-        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center" Spacing="20">
-            <Button Content="New Game" Style="{StaticResource MenuButtons}" Click="NewGame_Click"/>
-            <Button Content="Leaderboard" Style="{StaticResource MenuButtons}" Click="Leaderboard_Click"/>
-            <Button Content="Options" Style="{StaticResource MenuButtons}" Click="Options_Click"/>
-            <Button Content="Credits" Style="{StaticResource MenuButtons}" Click="Credits_Click"/>
-            <Button Content="Exit" Style="{StaticResource MenuButtons}" Click="Exit_Click"/>
-        </StackPanel>
-    </Grid>
-</Page>
